- Added 365 day or noleap calendar support (360 day calendar is on its way) - WCS supports all NetCDF datatypes - WCS _FillValue is now correctly set - GetFeatureInfo timeseries now works on all datatypes - Unsigned OpenDAP types are now detected correctly - e.g. A signed byte variable with _Unsigned attribute becomes unsigned byte variable.

+++ adagucserverEC/CImgRenderStippling.h
@@ -57,7 +57,7 @@ private:
     if(settings->hasNodataValue){if((val)==(T)settings->dfNodataValue)isNodata=true;}if(!(val==val))isNodata=true;
     if(!isNodata)if(settings->legendValueRange)if(val<settings->legendLowerRange||val>settings->legendUpperRange)isNodata=true;
     if(!isNodata){
-      if(x>=0&&y>=0&&x<settings->width&&y<settings->height){
+      if(x>=0&&y>=0&&x<(int)settings->width&&y<(int)settings->height){
         settings->dataField[x+y*settings->width]=(float)val; 
       }
     }
